refactor(ng-dev): separate out logic for takeover, target and regular checkouts
Separate the logic apart to be easier to parse as well as allow for intentional regular checkout.
